+++ CHANGELOG.rst
@@ -2,8 +2,16 @@ Changelog
 =========
 
 
+Version 2.0.0 (2018-02-11)
+--------------------------
 
-Version 1.1.0 (2018-01-24)
+- Nano rebrand - the `raiblocks` module has been replaced with `nano`
+- Pypi package name is now `nano-python`
+- Use `import nano` instead of `import raiblocks`
+- rpc client has been renamed from `nano.rpc.RPCClient` to `nano.rpc.Client`
+
+
+Version 1.1.0 (2018-02-11)
 --------------------------
 
 - RPC no longer automatically retries backend calls since this could lead to